HTML and BAT files for running the graphical interface (Biosignals Plux)
The HTML for graphically depicting the EEG, ECG and EDA readings (from the Biosignals Plux device) are included in the examples>> banana_demo folder.

A BAT file for running the program on the Anaconda environment is also included. In this file, the URL with the location of art-bio.py must be updated according to the local file address.